backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
